+++ src/backend/ci/core/quality/api-quality/src/main/kotlin/com/tencent/devops/quality/api/v2/UserQualityInterceptResource.kt
@@ -32,7 +32,8 @@ import com.tencent.devops.common.api.auth.AUTH_HEADER_USER_ID_DEFAULT_VALUE
 import com.tencent.devops.common.api.pojo.Page
 import com.tencent.devops.common.api.pojo.Result
 import com.tencent.devops.quality.pojo.RuleInterceptHistory
-import com.tencent.devops.quality.pojo.enum.RuleInterceptResult
+import com.tencent.devops.common.quality.pojo.enums.RuleInterceptResult
+import com.tencent.devops.common.quality.pojo.request.QualityHistoryListRequest
 import io.swagger.annotations.Api
 import io.swagger.annotations.ApiOperation
 import io.swagger.annotations.ApiParam
@@ -101,11 +102,8 @@ interface UserQualityInterceptResource {
         @ApiParam("构建ID", required = false)
         @QueryParam("buildId")
         buildId: String?,
-        @ApiParam("第几次检查", required = false)
-        @QueryParam("checkTimes")
-        checkTimes: Int?,
-        @ApiParam("规则ID", required = false)
-        ruleHashIds: Set<String>
+        @ApiParam("请求体", required = false)
+        request: List<QualityHistoryListRequest>
     ): Result<List<RuleInterceptHistory>>
 
     @ApiOperation("获取审核用户列表")
