+++ src/backend/ci/core/quality/api-quality/src/main/kotlin/com/tencent/devops/quality/api/v2/pojo/QualityIndicator.kt
@@ -30,7 +30,7 @@ package com.tencent.devops.quality.api.v2.pojo
 import com.tencent.devops.common.pipeline.pojo.element.agent.LinuxScriptElement
 import com.tencent.devops.common.pipeline.pojo.element.agent.WindowsScriptElement
 import com.tencent.devops.quality.api.v2.pojo.enums.QualityDataType
-import com.tencent.devops.quality.api.v2.pojo.enums.QualityOperation
+import com.tencent.devops.common.quality.pojo.enums.QualityOperation
 
 data class QualityIndicator(
     val hashId: String,
