+++ src/backend/ci/core/quality/api-quality/src/main/kotlin/com/tencent/devops/quality/api/v2/pojo/request/IndicatorCreate.kt
@@ -29,7 +29,7 @@ package com.tencent.devops.quality.api.v2.pojo.request
 
 import com.tencent.devops.quality.api.v2.pojo.enums.QualityDataType
 import io.swagger.annotations.ApiModel
-import com.tencent.devops.quality.api.v2.pojo.enums.QualityOperation
+import com.tencent.devops.common.quality.pojo.enums.QualityOperation
 
 @ApiModel("创建指标请求")
 data class IndicatorCreate(
