+++ src/backend/ci/core/quality/api-quality/src/main/kotlin/com/tencent/devops/quality/api/v2/pojo/response/IndicatorListResponse.kt
@@ -28,7 +28,7 @@
 package com.tencent.devops.quality.api.v2.pojo.response
 
 import com.tencent.devops.quality.api.v2.pojo.enums.QualityDataType
-import com.tencent.devops.quality.api.v2.pojo.enums.QualityOperation
+import com.tencent.devops.common.quality.pojo.enums.QualityOperation
 import io.swagger.annotations.ApiModel
 import io.swagger.annotations.ApiModelProperty
