+++ src/backend/ci/core/quality/biz-quality/src/main/kotlin/com/tencent/devops/quality/resources/v2/UserQualityInterceptResourceImpl.kt
@@ -31,10 +31,11 @@ import com.tencent.devops.common.api.exception.ParamBlankException
 import com.tencent.devops.common.api.pojo.Page
 import com.tencent.devops.common.api.pojo.Result
 import com.tencent.devops.common.api.util.PageUtil
+import com.tencent.devops.common.quality.pojo.enums.RuleInterceptResult
+import com.tencent.devops.common.quality.pojo.request.QualityHistoryListRequest
 import com.tencent.devops.common.web.RestResource
 import com.tencent.devops.quality.api.v2.UserQualityInterceptResource
 import com.tencent.devops.quality.pojo.RuleInterceptHistory
-import com.tencent.devops.quality.pojo.enum.RuleInterceptResult
 import com.tencent.devops.quality.service.v2.QualityHistoryService
 import com.tencent.devops.quality.service.v2.QualityRuleCheckService
 import org.springframework.beans.factory.annotation.Autowired
@@ -83,12 +84,9 @@ class UserQualityInterceptResourceImpl @Autowired constructor(
         projectId: String,
         pipelineId: String?,
         buildId: String?,
-        checkTimes: Int?,
-        ruleHashIds: Set<String>
+        request: List<QualityHistoryListRequest>
     ): Result<List<RuleInterceptHistory>> {
-        return Result(historyService.listInterceptHistoryForBuildHis(
-            userId, projectId, pipelineId, buildId, checkTimes, ruleHashIds)
-        )
+        return Result(historyService.listInterceptHistoryForBuildHis(userId, projectId, pipelineId, buildId, request))
     }
 
     override fun getAuditUserList(
